Move `StaticExtension` to public
