+++ src/output/get-loader-output.js
@@ -1,5 +1,5 @@
 import {TEMPLATE_DEPENDENCIES} from '../constants';
-import {ASSETS_KEY} from '../static-extension/contants';
+import {ASSETS_KEY} from '../public/static-extension/contants';
 
 import {getModuleOutput} from './get-module-output';
