+++ src/precompile/get-addons-meta.js
@@ -1,4 +1,4 @@
-import {getModule} from '../utils/get-module';
+import {getModule} from '../public/utils/get-module';
 
 /**
  * @param {[string, string]} addonEntry
