+++ src/public/static-extension/StaticExtension.js
