+++ src/public/static-extension/get-static-extension.js
